// desktop/src/open-api/model/upsertCustomFieldCommand.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UpsertCustomFieldOptionCommand } from './upsertCustomFieldOptionCommand';
import { CustomFieldType } from './customFieldType';


export interface UpsertCustomFieldCommand { 
    /**
     * Custom Field name
     */
    name: string;
    type: CustomFieldType;
    /**
     * Custom Field description
     */
    description?: string;
    options?: Array<UpsertCustomFieldOptionCommand>;
}
export namespace UpsertCustomFieldCommand {
}